* 各種表記をNinjaSlayer2_0からNinjaSlayer2に変更 * privateに関する表記を修正 * NinjaSlayerから継承していた処理をNinjaSlayer2内で実装
