+++ i18n/NinjaSlayer2/ja_jp.yml
@@ -1,0 +1,34 @@
+ja_jp:
+  NinjaSlayer2:
+    table:
+      SATSUBATSU:
+        name: "サツバツ!!(D6) ＞ "
+        type: 1D6
+        items:
+        - "「イヤーッ！」腹部に強烈な一撃が命中！ 敵はくの字に折れ曲がり、ワイヤーアクションめいて吹っ飛んだ！ \n『痛打+1』。敵の【体力】を減らした場合、付属効果として『弾き飛ばし』を与える。"
+        - "「観念してハイクを詠め！」頭部への痛烈なカラテが命中！ 眼球破壊もしくは激しい脳震盪が敵を襲う！ \n『痛打+1』。敵の【体力】を減らした場合、付属効果として『ニューロンダメージ2』と『ワザマエダメージ1』と『●部位損傷：頭部』を与える。"
+        - "「苦しみ抜いて死ぬがいい！」急所や内臓を情け容赦なく破壊！ \n『痛打+1』。敵の【体力】を減らした場合、付属効果として『ニューロンダメージ1』と『精神力ダメージ2』と『●部位損傷：胴体』を与える。"
+        - "「どこへ逃げても無駄だ！」敵の脚を無慈悲に粉砕！ \n『痛打+1』。敵の【体力】を減らした場合、付属効果として『カラテダメージ1』と『脚力ダメージ2』と『●部位損傷：脚部』を与える。"
+        - "「これで手も足も出まい！」敵の両腕をダブルチョップ切断！ 傷口から鮮血がスプリンクラーめいて噴き出す！ \n『痛打+1』。敵の【体力】を減らした場合、付属効果として『カラテダメージ2』と『ワザマエダメージ2』と『●部位損傷：腕部』を与える。"
+        - "「さらばだ！ イイイヤアアアアーーーーッ！」ヤリめいたチョップが敵の胸を貫通！ さらに心臓を掴み取り、握りつぶした！ ゴウランガ！ \n『即死！』。敵が『即死耐性』を持つ場合、この効果は『痛打＋2D6』に置き換えられる。"
+      WASSHOI:
+        name: "ニンジャスレイヤー=サンのエントリー!!(D6) ＞ "
+        type: 1D6
+        items:
+        - "高所からの回転着地！ タタミ四枚の距離で睨み合った！ \n標的ニンジャから3または4マス離れた任意のマスに、【殺】コマを置くこと。"
+        - "ドアを蹴破って出現！ \n標的ニンジャがいる部屋の任意のドアの隣に【殺】コマを置くこと。 \nなお、ドアにあらかじめ鍵をかけるなどの行為は全て無駄である。"
+        - "KRAAAAASH！ 窓を突き破り出現！ \n標的ニンジャがいる部屋の任意の窓の隣に【殺】コマを置くこと。 \n【殺】コマが隣接している間、その窓は脱出用として使用できなくなる。"
+        - "天井破壊や床破砕、または垂直リフト射出により出現！ \n標的ニンジャから2マス離れた任意の場所に【殺】コマを置くこと。 \n激しい恐怖や動揺により、次のターンの終了時まで、その場にいる【DKK】1以上のニンジャ全員は『連続側転判定』の難易度が+1される。"
+        - "冷蔵庫や金庫から突如出現！ \nそのマップ上に存在するトレジャーボックス内に、ニンジャスレイヤーが潜んでいた。標的ニンジャから最も近いトレジャーボックス1個（もしくは適切な障害物や爆発物）の隣に【殺】コマを置くこと。 \n激しい恐怖や動揺により、次のターンの終了時まで、その場にいる【DKK】1以上のニンジャ全員は『連続側転判定』の難易度が+1される。"
+        - "「行き先はジゴクですよ」 \nマップ上にいるNPC1人（標的ニンジャから最も近くにいる者）が、実はニンジャスレイヤーの変装であった。そのNPCのコマを【殺】に変更せよ（本物のNPCがどこにいったのかはニンジャマスターが後で考える）。 \n激しい恐怖や動揺により、次のターンの終了時まで、その場にいる【DKK】1以上のニンジャ全員は『連続側転判定』の難易度が+2される。"
+      NRS:
+        name: "ニンジャ・リアリティ・ショック発症(D7) ＞ "
+        type: 1D7
+        items:
+        - "絶叫 \n恐怖のあまり、その場で絶叫を上げる。 \n探索中の場合、近くの敵に存在を気づかれる可能性がある（シナリオ次第）。 \n戦闘中の場合、恐怖のあまり足がガクガクと震え、このターン終了まで【脚力】が0となり、『崩れ状態』とみなされる。"
+        - "失禁 \n恐怖とトラウマのあまりその場で立ちすくみ、失禁する。自尊心を失い【精神力】が-1される。 \n戦闘中の場合、足がすくんで手がブルブルと震え、このターン終了まであらゆる自発的行動の難易度が＋１され、【脚力】が１となり、『崩れ状態』とみなされる。"
+        - "パニック逃走や異常行動 \n恐怖のあまり絶叫し、仲間を見捨ててその場から全速力で逃げ出そうとしたり、銃を持っている場合は見えない敵に向かって乱射しようとする。 \nあるいは仲間のことをニンジャだと思い込んで攻撃を仕掛けたり、理解不能な異常行動を取ったりする。 \n探索シーケンスの場合、直ちに戦闘シーケンスが発生してこのPCの手番となり、1ターンだけマスターがこのPCを操作する。このPCは常軌を逸した身体能力を発揮し、【脚力】6として行動する。マスターはこのPCに自傷行為以外のどんな行動を取らせてもよい。このPCにとっては、あらゆるキャラが敵とみなされる。このターンの終了時に、このPCは正気に戻る。 \n戦闘中の場合も同様で、次の手番にこの状態となる。手番開始時に攻撃できる相手がニンジャしか見えていない場合、逃走を優先する。"
+        - "ドゲザ \n失禁し【精神力】に1ダメージ。突然ドゲザによる命乞いを行うため、このターン終了時まで『麻痺状態』とみなされる。"
+        - "激しい嘔吐や鼻血 \n【体力】と【精神力】にそれぞれ１ダメージを受ける。 \n戦闘中の場合、このターンのあらゆる判定難易度が+1される。 \nしかし目の前の脅威に立ち向かおうとする意志は砕けていない。"
+        - "気絶 \n失禁し【精神力】に1ダメージ。さらに打ち上げられたマグロめいてその場で倒れて口をパクパクとさせ、『気絶状態』となる。 \n戦闘が終了するか、一定時間が経過するか、誰かに蘇生してもらうまで、この状態は解除されない。"
+        - "心臓発作や狂死 \n急激なニンジャリアリティショックに耐えきれず、PCは心臓発作やニューロン損傷を起こして【体力】0となり、その場に倒れ『気絶状態』となる。 \nZBRアドレナリンでなければ蘇生できない。処置を受けずに時間が経過すると死亡する。"
