+++ lib/bcdice/game_system.rb
@@ -152,7 +152,7 @@ require "bcdice/game_system/NightWizard"
 require "bcdice/game_system/NightWizard3rd"
 require "bcdice/game_system/NightmareHunterDeep"
 require "bcdice/game_system/NinjaSlayer"
-require "bcdice/game_system/NinjaSlayer2_0"
+require "bcdice/game_system/NinjaSlayer2"
 require "bcdice/game_system/NjslyrBattle"
 require "bcdice/game_system/NobunagasBlackCastle"
 require "bcdice/game_system/NSSQ"
